feat: Add test for SVG script injection (CVE-2021-43818) (#53)
Adds a new security test to ensure that the XML parser is not vulnerable to script injection via SVG images embedded in data URIs (CVE-2021-43818).

Although the project already uses a version of `lxml` that is not vulnerable, this test provides an extra layer of defense to prevent future regressions.